today i complete the full project with finishing bookings api

+++ src/module/users/user.service.ts
@@ -1,15 +1,30 @@
 import { pool } from "../../database/db";
 
-// Get all vehicles
+// Get all user
 const getAllUsers = async () => {
   const result = await pool.query(`
-    SELECT * FROM users`);
+    SELECT id, name,email, phone, role FROM users`);
   return result;
 };
 
-//update vehicle
-const updateUser = async (payload: Record<string, unknown>, id: string) => {
+//update user
+const updateUser = async (
+  payload: Record<string, unknown>,
+  id: string,
+  currentUser: Record<string, any>
+) => {
   const { name, email, password, phone, role } = payload;
+
+  if (currentUser.role !== "admin") {
+    if (currentUser.id !== Number(id)) {
+      throw new Error("You can update only your own profile");
+    }
+
+    if (role) {
+      throw new Error("You are not allowed to change role");
+    }
+  }
+
   const result = await pool.query(
     `
     UPDATE users
@@ -27,8 +42,16 @@ const updateUser = async (payload: Record<string, unknown>, id: string) => {
   return result;
 };
 
-//delete vehicle
+//delete user
 const deleteUser = async (id: string) => {
+  const booking = await pool.query(
+    `SELECT id FROM bookings WHERE customer_id=$1`,
+    [id]
+  );
+
+  if (booking.rowCount !== 0) {
+    throw new Error("user not delete");
+  }
   const result = await pool.query(
     `
     DELETE FROM users WHERE id=$1
